# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
      "SessionBody": {
        "additionalProperties": false,
        "properties": {
          "state": {
            "enum": [
              "anonymous",
              "pendingMfa",
              "authenticated"
            ],
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/SessionUserBody"
          }
        },
        "required": [
          "state"
        ],
        "type": "object"
      },
      "SessionUserBody": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "role": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "role"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "Encrypted session cookie, served as __Host-step-ui when SESSION_SECURE is on (D6). Role names in a requirement array are documentation: enforcement is server-side.",
        "in": "cookie",
        "name": "step-ui",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "step-ca-ui",
    "version": "1.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/session": {
      "delete": {
        "operationId": "deleteSession",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [],
        "summary": "Log out and expire the session and CSRF cookies",
        "tags": [
          "session"
        ]
      },
      "get": {
        "operationId": "getSession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [],
        "summary": "Get the current session state",
        "tags": [
          "session"
        ]
      }
    },
...
  }
}